Refactor for better code quality

+++ app/src/main/java/com/example/icard_ex/DetailsActivity.kt
@@ -1,5 +1,6 @@
 package com.example.icard_ex
 
+import android.app.AlertDialog
 import android.content.Intent
 import android.os.Bundle
 import android.view.Menu
@@ -12,9 +13,9 @@ import kotlinx.android.synthetic.main.activity_details.*
 class DetailsActivity : AppCompatActivity() {
     private lateinit var dbHelper       :   DatabaseHelper
     private lateinit var contactsHelper :   ContactsHelper
+    private lateinit var contact        :   Contact
     private var id                      =   -1
     private var idDevice                =   0
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_details)
@@ -40,8 +41,13 @@ class DetailsActivity : AppCompatActivity() {
         var isOptionSelected = true
         when (item.itemId) {
             R.id.actionDelete   ->  {
-                dbHelper.deleteContact(id)
-                finish()
+                AlertDialog.Builder(this).setMessage("Delete ${contact.fullname}?")
+                    .setPositiveButton("Yes") { _, _ ->
+                        dbHelper.deleteContact(id)
+                        finish()
+                    }
+                    .setNegativeButton("No"){_, _ ->
+                    }.show()
             }
             R.id.actionEdit     ->  {
                 if (id == defaultID)
@@ -69,7 +75,6 @@ class DetailsActivity : AppCompatActivity() {
     }
 
     private fun setUpDetails(){
-        val contact :   Contact
         val country :   Country
         if(id != defaultID) {
             contact             =   dbHelper.getContact(id)
